refactor: address all code review findings from simplifier and perf agents
- Extract buildActivityData to lib/activity.ts (decouples page from API route)
- Move buildHiroHeaders and detect429 to lib/stacks-api-fetch.ts (shared by 6 files)
- Remove duplicate NetworkStats interface from activity route (use shared type)
- Fix setQueuedStats anti-pattern: initialize from computed value, not inside
  another useState initializer
- Simplify negative txid cache writes from structured JSON to "1" (only
  existence is checked on read)
- Remove formatNumber wrapper, inline toLocaleString() at call sites
- Trim unused LeaderboardAgent fields (rank, level, levelName)
- Document inbox:pending-txid:* KV key pattern in CLAUDE.md

Net -200 lines.
