+++ lib/stacks-api-fetch.ts
@@ -13,6 +13,35 @@
  * - Returns the final Response after all retries — callers check status
  */
 
+/** Build headers for Hiro API requests, optionally including an API key. */
+export function buildHiroHeaders(hiroApiKey?: string): Record<string, string> {
+  const headers: Record<string, string> = {};
+  if (hiroApiKey) {
+    headers["X-Hiro-API-Key"] = hiroApiKey;
+  }
+  return headers;
+}
+
+/**
+ * Detect 429 rate limit responses and log cf-ray for observability.
+ *
+ * @returns Object with isRateLimited flag for caller branching
+ */
+export function detect429(response: Response): {
+  isRateLimited: boolean;
+} {
+  const isRateLimited = response.status === 429;
+
+  if (isRateLimited) {
+    const cfRay = response.headers.get("cf-ray");
+    console.warn(
+      `Rate limit detected (429) on ${response.url}${cfRay ? ` [cf-ray: ${cfRay}]` : ""}`
+    );
+  }
+
+  return { isRateLimited };
+}
+
 /** Per-attempt fetch timeout in milliseconds. */
 const PER_ATTEMPT_TIMEOUT_MS = 8_000;
